Canopy-Metrics-main/Canopy-Metrics-main/requirements.txt
streamlit>=1.28
ultralytics>=8.0.206
opencv-python>=4.8
numpy>=1.24
Pillow>=10.0
# Use a newer PyTorch that supports Python 3.12+
torch>=2.2
torchvision>=0.17
